chore: refactored sidebar
