+++ src/app/app.html
@@ -6,9 +6,9 @@
   ></spartan-menubar>
 
   <div class="app-layout flex-1 flex overflow-hidden">
-    <ctf-chat-sidebar
+    <ctf-workspace-sidebar
       class="flex-none border-r border-border transition-[width] duration-300 ease-in-out"
-    ></ctf-chat-sidebar>
+    ></ctf-workspace-sidebar>
 
     <main class="app-main flex-1 overflow-auto relative">
       <router-outlet></router-outlet>
